test(ci): protect playwright workflow coverage
